+++ RandomEncounters/Encounter/Flotsam.cs
@@ -7,33 +7,33 @@ namespace RandomEncounters
     internal class Flotsam
     {
         //cargos
-        //"1 crate salmon (E)",
-        //"2 crate dates (good)",
-        //"3 crate coconuts (good)",
-        //"4 crate lamb (good)",
-        //"5 crate tea (good)",
-        //"6 crate tuna (A)",
-        //"7 crate cheese (good)",
-        //"8 crate goat cheese (good)",
-        //"9 crate sunspot fish (A)",
-        //"10 barrel water",
-        //"11 barrel rum",
-        //"12 barrel beer" // for some reason look at text says barrel of water,
-        //"13 barrel wine",
-        //"14 crate north fish (M)",
-        //"15 crate sausages",
-        //"16 crate pork",
-        //"17 crate bananas",
-        //"18 crate trout (M)",
-        //"19 crate eel (E)",
-        //"24 barrel spices",
-        //"25 crate grain",
-        //"26 crate medicine",
-        //"27 crate seafood",
+        //1 crate salmon (E)
+        //2 crate dates (good)
+        //3 crate coconuts (good)
+        //4 crate lamb (good)
+        //5 crate tea (good)
+        //6 crate tuna (A)
+        //7 crate cheese (good)
+        //8 crate goat cheese (good)
+        //9 crate sunspot fish (A)
+        //10 barrel water
+        //11 barrel rum
+        //12 barrel beer // look text says barrel of water,
+        //13 barrel wine
+        //14 crate north fish (M)
+        //15 crate sausages
+        //16 crate pork
+        //17 crate bananas
+        //18 crate trout (M)
+        //19 crate eel (E)
+        //24 barrel spices
+        //25 crate grain        // too heavy
+        //26 crate medicine
+        //27 crate seafood
         //201 crate venison
         //202 crate truffles
         //206 barrel mead
-        //212 crate rice
+        //212 crate rice        // too heavy
         //213 crate oranges
         //214 crate forest mushrooms
         //216 crate cave mushrooms
@@ -44,37 +44,52 @@ namespace RandomEncounters
         //224 apples
         //227 sulfur
         //228 barrel cider
+        //229 hemp
+        //230 dyes
+        //231 rubber
+        //232 coffee
+        //233 salt
+        //234 saffron
 
         //consumables
-        //"104 crate of fishing hooks",
-        //"108 crate of firewood",
-        //"131 lantern candle crate",
-        //"132 lantern oil bottle"
+        //104 crate of fishing hooks
+        //108 crate of firewood
+        //131 lantern candle crate
+        //132 lantern oil bottle
 
         //bottles
-        //"55 water bottle",
-        //"56 coco wine",
-        //"57 honey beer",
-        //"58 rice beer",
-        //"59 wine"
+        //55 water bottle
+        //56 coco wine
+        //57 honey beer
+        //58 rice beer
+        //59 wine
 
         //tobaccos
-        //"311 crate of tobacco white",
-        //"313 crate of tobacco green",
-        //"315 crate of tobacco black",
-        //"317 crate of tobacco brown",
-        //"319 crate of tobacco blue"
+        //311 crate of tobacco white
+        //313 crate of tobacco green
+        //315 crate of tobacco black
+        //317 crate of tobacco brown
+        //319 crate of tobacco blue
+
+        //tea and coffee
+        //387 tea box white
+        //388 tea box black
+        //389 tea box green
+        //373 coffee box
+        //386 coffee barrel
+
 
         private static readonly int[] _cargos = 
         { 
             1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 
-            11, 13, 14, 15, 16, 17, 18, 19, 24, 25,
-            26, 27, 201, 202, 206, 212, 213, 214, 216, 219,
-            220, 222, 223, 224, 227, 228,
+            11, 13, 14, 15, 16, 17, 18, 19, 24, 26,
+            27, 201, 202, 206, 213, 214, 216, 219, 220, 222,
+            223, 224, 227, 228, 229, 230, 231, 232, 233, 234
         };
         private static readonly int[] _consumables = { 104, 108, 131, 132 };
         private static readonly int[] _bottles = { 55, 56, 57, 58, 59 };
         private static readonly int[] _tobaccoCrates = { 311, 313, 315, 319 };
+        private static readonly int[] _teaAndCoffeeBoxes = { 387, 388, 389, 373 };
 
         private const int MAX_CARGO_TYPES = 4;
         private const int MAX_CARGOS = 3;
@@ -88,7 +103,7 @@ namespace RandomEncounters
 
                 for (int j = 0; j < Random.Range(1, MAX_CARGOS); j++)
                 {
-                    LogDebug($"Choice: {_cargos[choice]}");
+                    LogDebug($"Cargo choice: {_cargos[choice]}");
                     var prefabGO = PrefabsDirectory.instance.directory[_cargos[choice]];
                     var amount = (float)System.Math.Round((decimal)Random.Range(0, prefabGO.GetComponent<ShipItem>().amount));
                     SpawnItem(spawnPoint, prefabGO, amount);
@@ -98,7 +113,7 @@ namespace RandomEncounters
             for (int i = 0; i < Random.Range(1, MAX_CONSUME_TYPES); i++)
             {
                 var choice = Random.Range(0, _consumables.Length - 1);
-                LogDebug($"Choice: {_consumables[choice]}");
+                LogDebug($"Consumable choice: {_consumables[choice]}");
                 var prefabGO = PrefabsDirectory.instance.directory[_consumables[choice]];
                 var amount = (float)System.Math.Round((decimal)Random.Range(0, prefabGO.GetComponent<ShipItem>().amount));
                 SpawnItem(spawnPoint, prefabGO, amount);
@@ -107,17 +122,23 @@ namespace RandomEncounters
             for (int i = 0; i < Random.Range(5, 10); i++)
             {
                 var choice = Random.Range(0, _bottles.Length - 1);
-                LogDebug($"Choice: {_bottles[choice]}");
+                LogDebug($"Bottle choice: {_bottles[choice]}");
                 var prefabGO = PrefabsDirectory.instance.directory[_bottles[choice]];
                 var amount = 0f;
                 SpawnItem(spawnPoint, prefabGO, amount);
             }
 
             var tobaccoChoice = Random.Range(0, _tobaccoCrates.Length - 1);
-            LogDebug($"Choice: {_tobaccoCrates[tobaccoChoice]}");
+            LogDebug($"Tobacco choice: {_tobaccoCrates[tobaccoChoice]}");
             var tobaccoPrefabGO = PrefabsDirectory.instance.directory[_tobaccoCrates[tobaccoChoice]];
             var tobaccoAmount = (float)System.Math.Round((decimal)Random.Range(0, tobaccoPrefabGO.GetComponent<ShipItem>().amount));
             SpawnItem(spawnPoint, tobaccoPrefabGO, tobaccoAmount);
+
+            var teaCoffeeChoice = Random.Range(0, _teaAndCoffeeBoxes.Length - 1);
+            LogDebug($"Tea/Coffee choice: {_teaAndCoffeeBoxes[teaCoffeeChoice]}");
+            var teaCoffeePrefabGO = PrefabsDirectory.instance.directory[_teaAndCoffeeBoxes[teaCoffeeChoice]];
+            var teaCoffeeAmount = (float)System.Math.Round((decimal)Random.Range(0, teaCoffeePrefabGO.GetComponent<ShipItem>().amount));
+            SpawnItem(spawnPoint, teaCoffeePrefabGO, teaCoffeeAmount);
 
             SpawnItem(spawnPoint, AssetLoader.SmallWreck, 1f, true);
         }
@@ -154,7 +175,7 @@ namespace RandomEncounters
             for (int i = 0; i < num; i++)
             {
                 LogDebug("Inserting item " + crate.amount);
-                GameObject gameObject = Object.Instantiate(crate.GetContainedPrefab(), crate.transform.position + new Vector3(0f, 100.5f, 0f), crate.transform.rotation);
+                var gameObject = Object.Instantiate(crate.GetContainedPrefab(), crate.transform.position + new Vector3(0f, 100.5f, 0f), crate.transform.rotation);
                 crate.amount -= 1f;
                 gameObject.GetComponent<SaveablePrefab>().RegisterToSave();
                 if ((bool)gameObject.GetComponent<CookableFood>())